[BOOKINGSG-6090][JZ] Remove and rename files
